feat(ui): refactor set username modal
