+++ src/ui/App.tsx
@@ -9,7 +9,7 @@ import {
 import { IonReactRouter } from "@ionic/react-router";
 import { StrictMode, useEffect, useState } from "react";
 import { RoutePath, Routes } from "../routes";
-import { PublicRoutes, TabsRoutePath } from "../routes/paths";
+import { PublicRoutes } from "../routes/paths";
 import { useAppSelector } from "../store/hooks";
 import {
   getAuthentication,
@@ -18,12 +18,11 @@ import {
   getIsInitialized,
   getIsOnline,
 } from "../store/reducers/stateCache";
-import "./App.scss";
 import { AppOffline } from "./components/AppOffline";
 import { AppWrapper } from "./components/AppWrapper";
 import { ToastStack } from "./components/CustomToast/ToastStack";
 import { GenericError } from "./components/Error";
-import { SetUserName } from "./components/SetUserName";
+import { InputModal } from "./components/InputModal";
 import { SidePage } from "./components/SidePage";
 import { OperationType } from "./globals/types";
 import { FullPageScanner } from "./pages/FullPageScanner";
@@ -31,6 +30,7 @@ import { LoadingPage } from "./pages/LoadingPage/LoadingPage";
 import { LockPage } from "./pages/LockPage/LockPage";
 import "./styles/ionic.scss";
 import "./styles/style.scss";
+import "./App.scss";
 
 setupIonicReact();
 
@@ -39,7 +39,6 @@ const App = () => {
   const isOnline = useAppSelector(getIsOnline);
   const authentication = useAppSelector(getAuthentication);
   const currentRoute = useAppSelector(getCurrentRoute);
-  const [showSetUserName, setShowSetUserName] = useState(false);
   const currentOperation = useAppSelector(getCurrentOperation);
   const [showScan, setShowScan] = useState(false);
 
@@ -55,17 +54,6 @@ const App = () => {
       ].includes(currentOperation)
     );
   }, [currentOperation]);
-
-  useEffect(() => {
-    if (
-      authentication.loggedIn &&
-      (authentication.userName === undefined ||
-        authentication.userName?.length === 0) &&
-      currentRoute?.path?.includes(TabsRoutePath.ROOT)
-    ) {
-      setShowSetUserName(true);
-    }
-  }, [authentication.loggedIn, authentication.userName, currentRoute]);
 
   useEffect(() => {
     const platforms = getPlatforms();
@@ -123,10 +111,7 @@ const App = () => {
           ) : (
             <LoadingPage />
           )}
-          <SetUserName
-            isOpen={showSetUserName}
-            setIsOpen={setShowSetUserName}
-          />
+          <InputModal />
           <SidePage />
           <GenericError />
           <ToastStack />
