+++ src/ui/components/InputModal/InputModal.scss
@@ -1,17 +1,25 @@
-ion-modal#set-connection-alias {
-  --backdrop-opacity: var(--ion-backdrop-opacity, 0.4);
+ion-modal#input-modal {
   padding: 1.25rem;
   --height: fit-content;
   --border-radius: 1rem;
   --box-shadow: 2px 2px 10px 0px rgba(3, 3, 33, 0.3);
 
-  .set-connection-alias-wrapper {
+  &.connection-alias {
+    --backdrop-opacity: var(--ion-backdrop-opacity, 0.4);
+
+    .input-modal-wrapper {
+      h3 {
+        font-size: 1.125rem;
+      }
+    }
+  }
+
+  .input-modal-wrapper {
     padding: 1.56rem 1.25rem;
 
     h3 {
       font-weight: 500;
       text-align: center;
-      font-size: 1.125rem;
       line-height: 1.25rem;
     }
 
@@ -30,7 +38,7 @@ ion-modal#set-connection-alias {
   }
 
   @media screen and (min-width: 250px) and (max-width: 399px) {
-    .set-connection-alias-wrapper {
+    .input-modal-wrapper {
       padding: 1.248rem 1rem;
 
       h3 {
