+++ src/ui/components/InputModal/index.ts
@@ -1,0 +1,1 @@
+export * from "./InputModal";
