--- src/ui/components/SetConnectionAlias/SetConnectionAlias.types.ts
@@ -1,7 +1,0 @@
-interface SetConnectionAliasProps {
-  isOpen: boolean;
-  setIsOpen: (value: boolean) => void;
-  onConfirm: (name: string) => void;
-}
-
-export type { SetConnectionAliasProps };
