--- src/ui/components/SetConnectionAlias/index.ts
@@ -1,1 +1,0 @@
-export * from "./SetConnectionAlias";
