--- src/ui/components/SetUserName/SetUserName.scss
@@ -1,50 +1,0 @@
-ion-modal#set-user-name {
-  padding: 1.25rem;
-  --height: fit-content;
-  --border-radius: 1rem;
-  --box-shadow: 2px 2px 10px 0px rgba(3, 3, 33, 0.3);
-
-  .set-user-name-wrapper {
-    padding: 1.56rem 1.25rem;
-
-    h3 {
-      font-weight: 500;
-      text-align: center;
-    }
-
-    ion-item.custom-input {
-      --background: white;
-      margin: 1.563rem 0;
-
-      ion-label {
-        font-size: 1rem;
-      }
-    }
-
-    ion-toolbar.page-footer {
-      padding: 0;
-    }
-  }
-
-  @media screen and (min-width: 250px) and (max-width: 399px) {
-    .set-user-name-wrapper {
-      padding: 1.248rem 1rem;
-
-      h3 {
-        font-size: 0.95rem;
-      }
-
-      ion-item.custom-input {
-        margin: 1.25rem 0;
-
-        ion-label {
-          font-size: 0.8rem;
-        }
-      }
-    }
-  }
-
-  @media screen and (min-width: 250px) and (max-width: 374px) {
-    padding: 1rem;
-  }
-}
