--- src/ui/components/SetUserName/SetUserName.types.ts
@@ -1,6 +1,0 @@
-interface SetUserNameProps {
-  isOpen: boolean;
-  setIsOpen: (value: boolean) => void;
-}
-
-export type { SetUserNameProps };
