--- src/ui/components/SetUserName/index.ts
@@ -1,1 +1,0 @@
-export * from "./SetUserName";
