+++ src/ui/pages/Connections/Connections.tsx
@@ -52,7 +52,6 @@ import { ScrollablePageLayout } from "../../components/layout/ScrollablePageLayo
 import { PageHeader } from "../../components/PageHeader";
 import { ConnectionDetails } from "../ConnectionDetails";
 import { CreateIdentifier } from "../../components/CreateIdentifier";
-import { SetConnectionAlias } from "../../components/SetConnectionAlias";
 
 const Connections = forwardRef<ConnectionsOptionRef, ConnectionsComponentProps>(
   ({ showConnections, setShowConnections }, ref) => {
@@ -350,7 +349,6 @@ const Connections = forwardRef<ConnectionsOptionRef, ConnectionsComponentProps>(
           )}`}
           onDeletePendingItem={deleteConnection}
         />
-        <SetConnectionAlias />
       </>
     );
   }
